requirements.txt
streamlit>=1.28.0
psycopg2-binary
pandas
python-dotenv
chromadb
google-generativeai>=0.3.0
pysqlite3-binary
numpy
langchain-community
altair